use template and environment variables for Nextcloud config
The NextCloud configuration file is now based on a template, rather than
attempting to hack the auto-generated config output by the installation.

We capture the generated values from the installer-generated config and use
EnvPlate to replace the values in our template. We then use environment
variables for those settings it makes sense to configure dynamically. This
includes mail settings, but excludes DB settings which are set during the
install and should never need to be changed afterwards.