use a single Docker volume for NextCloud data
This consolidates all NextCloud data and config files into a single Docker
volume location, /var/lib/nextcloud, which is where the single nextcloud_data
Docker volume should be mounted.

The base image already defines five volumes, which means we have to copy the
files out of these locations into our volume at run time. The exception to this
is the /nextcloud/themes volume, which cannot be moved because NextCloud appears
to have this path hard-coded so can't be reconfigured to use a different path,
therefore no amount of copying or symlinking can change where it is located.

The arkivum-run.sh has been renamed to run.sh and the content of the upstream
run.sh has been copied into it and modified. The changes required are too much
to "bolt on" to the upstream script, so instead it's been replaced entirely.

The arkivum-setup.sh has been changed to use the new path, and the config.php
template has now been moved into the main nextcloud folder to avoid it being
masked by the volume mount.