make dbname config setting configurable via environment variable
The dbname value was previously hardcoded, however it is not as predictable
as previously thought so it needs to be set using environment variable. The
upstream Docker image actually already supported this, so this fixes the
expected behaviour.

I've also made dbhost and dbport more dynamic, so they are read each time
the instance runs rather than just at install time. The dbuser and
dbpassword are still install-only values, because the dbpassword has to be
the crypted value, not plaintext.