make redis hostname and port configurable using environment variables
We previously assumed that we always knew these, but this isn't the case.
They are now therefore configuable using REDIS_HOST and REDIS_PORT.

We don't make any provision for specifying multiple hosts in a cluster as
this is currently unnecessary. If a Redis cluster is to be used then it
should be done behind a single DNS hostname, so that the application has no
knowledge of it.