disable update checker in config
We want to disable NextCloud's update checker because if we want to update
the version of NextCloud being deployed then we will update the image and
then re-deploy the service with a new image. We don't want NextCloud or
users to update the software themselves, as doing so wipes our config and
resets things like external storage locations etc.

This fixes #11.