read version number from NextCloud code
We used to hardcode the version, but this broke when the upstream image was
updated. We now read the version from the nextcloud/version.php file, and
store the value in /var/lib/nextcloud/config/version. Each time the
instance is started we compare this with what is in the code and, if it's
changed, we update our config file.