avoid using wonderfall/nextcloud as base image
This upstream image has been discontinued, plus it uses s6 which is very
hard to configure and maintain.

We now base our image on boxedcode/alpine-nginx-php-fpm, which whilst not
a pure OS image is a useful compromise, including ready-built nginx and
PHP-FPM. It uses supervisor to run the multiple processes.

The nextcloud install and configuration from the wonderfall image has been
copied into our image, so we're now in full control of what nextcloud
version etc is installed and how it's configured. It also means we can
remove a number of shims that were in place to change the default
behaviour.