check if external storages exist before creating them
We now use the occ files_external:list command in nextcloud-setup to check
if a storage with the given name exists before attempting to create it.

Although the setup script is only intended to be run once, this protects
against it being run multiple times inadvertently, which would result in
duplicate external storages being created.