add user_saml app to install to enable shibboleth support
The user_saml app implements SAML-based SSO authentication and
authorisation, which means we can enable Shibboleth if required.

This is done using the SHIBBOLETH_ENABLED environment variable. If set,
the nextcloud-shibboleth script is run, which enables the app in NextCloud
and also injects the required config settings in the database from a
templated SQL script.

If SHIBBOLETH_ENABLED is set, the SHIB_IDP_CERT_FILE, SHIB_SP_CERT_FILE
and SHIB_SP_KEY_FILE should also be set, to the paths of each of the IdP
signing certificate, SP encrypt-and-sign certificate and encrypt-and-sign
private key.

We currently use a custom fork of the user_saml app, as we want to map
Shibboleth entitlements to NextCloud groups, which the main version doesn't
support. At some point this feature may be contributed back to the main
fork.